sensors/requirements.txt
pyyaml>=6.0
aiohttp>=3.9.0
